src/indexer/requirements.txt
scrapy==2.4.0
beautifulsoup4
pysolr
python-dateutil
psycopg2
tldextract

